refactor(api): move window config related functions to own file
